Safari-611.1.10/LayoutTests/fast/dom/domtimestamp-is-number.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body onload="do_check(event)">
<script>
description("This tests that DOMTimeStamp is a Number (and not a Date object).");
// See https://bugs.webkit.org/show_bug.cgi?id=49963

var timestamp = null;

function do_check(e) {
  timestamp = e.timeStamp;
  shouldBeFalse("timestamp instanceof Date");
  shouldBeTrue("timestamp == Number(timestamp)");
  finishJSTest();
}
window.jsTestIsAsync = true;</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
