third_party/blink/web_tests/platform/generic/fast/dom/domtimestamp-is-number-expected.txt

This tests that DOMTimeStamp is a Number (and not a Date object).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timestamp instanceof Date is false
PASS timestamp == Number(timestamp) is true
PASS successfullyParsed is true

TEST COMPLETE
